requirements.txt
# Web scraping and HTTP requests
requests>=2.31.0
beautifulsoup4>=4.12.0


# Environment variables
python-dotenv==1.0.0

# OpenAI integration for captcha recognition
openai>=1.0.0

sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL适配器

pytest